requirements.txt
django==2.1.4
djangorestframework==3.9.0
jsonfield